requirements.txt
paddlepaddle==2.6.0
paddleocr==2.7.0.3
flask
gunicorn
numpy
pillow